De-cluttered P/Invoke code
