+++ src/Files.App/Helpers/Interop/IDataTransferManagerInterop.cs
@@ -1,0 +1,17 @@
+// Copyright (c) 2023 Files Community
+// Licensed under the MIT License. See the LICENSE.
+
+using System.Runtime.InteropServices;
+
+namespace Files.App.Helpers
+{
+	[ComImport]
+	[Guid("3A3DCD6C-3EAB-43DC-BCDE-45671CE800C8")]
+	[InterfaceType(ComInterfaceType.InterfaceIsIUnknown)]
+	interface IDataTransferManagerInterop
+	{
+		IntPtr GetForWindow([In] IntPtr appWindow, [In] ref Guid riid);
+
+		void ShowShareUIForWindow(IntPtr appWindow);
+	}
+}
