+++ src/Files.App/Helpers/Layout/LayoutPreferencesManager.cs
@@ -499,7 +499,7 @@ namespace Files.App.Data.Models
 		{
 			if (!UserSettingsService.FoldersSettingsService.SyncFolderPreferencesAcrossDirectories)
 			{
-				var folderFRN = NativeFileOperationsHelper.GetFolderFRN(path);
+				var folderFRN = Win32PInvoke.GetFolderFRN(path);
 				var trimmedFolderPath = path.TrimPath();
 				if (trimmedFolderPath is not null)
 					SetLayoutPreferencesToDatabase(trimmedFolderPath, folderFRN, preferencesItem);
@@ -613,7 +613,7 @@ namespace Files.App.Data.Models
 			{
 				path = path.TrimPath() ?? string.Empty;
 
-				var folderFRN = NativeFileOperationsHelper.GetFolderFRN(path);
+				var folderFRN = Win32PInvoke.GetFolderFRN(path);
 
 				return GetLayoutPreferencesFromDatabase(path, folderFRN)
 					?? GetLayoutPreferencesFromAds(path, folderFRN)
@@ -625,7 +625,7 @@ namespace Files.App.Data.Models
 
 		private static LayoutPreferencesItem? GetLayoutPreferencesFromAds(string path, ulong? frn)
 		{
-			var str = NativeFileOperationsHelper.ReadStringFromFile($"{path}:files_layoutmode");
+			var str = Win32PInvoke.ReadStringFromFile($"{path}:files_layoutmode");
 
 			var layoutPreferences = SafetyExtensions.IgnoreExceptions(() =>
 				string.IsNullOrEmpty(str) ? null : JsonSerializer.Deserialize<LayoutPreferencesItem>(str));
@@ -635,7 +635,7 @@ namespace Files.App.Data.Models
 
 			// Port settings to the database, delete the ADS
 			SetLayoutPreferencesToDatabase(path, frn, layoutPreferences);
-			NativeFileOperationsHelper.DeleteFileFromApp($"{path}:files_layoutmode");
+			Win32PInvoke.DeleteFileFromApp($"{path}:files_layoutmode");
 
 			return layoutPreferences;
 		}
