+++ src/Files.App/Utils/Storage/Helpers/FilePropertiesHelpers.cs
@@ -133,7 +133,7 @@ namespace Files.App.Utils.Storage
 				new SuppressNavigationTransitionInfo());
 
 			// WINUI3: Move window to cursor position
-			InteropHelpers.GetCursorPos(out var pointerPosition);
+			Win32PInvoke.GetCursorPos(out var pointerPosition);
 			var displayArea = DisplayArea.GetFromPoint(new PointInt32(pointerPosition.X, pointerPosition.Y), DisplayAreaFallback.Nearest);
 			var appWindowPos = new PointInt32
 			{
