+++ src/Files.App/Utils/Storage/Operations/FileOperationsHelpers.cs
@@ -661,7 +661,7 @@ namespace Files.App.Utils.Storage
 
 		public static IEnumerable<Win32Process>? CheckFileInUse(string[] fileToCheckPath)
 		{
-			var processes = SafetyExtensions.IgnoreExceptions(() => FileUtils.WhoIsLocking(fileToCheckPath), App.Logger);
+			var processes = SafetyExtensions.IgnoreExceptions(() => Win32PInvoke.WhoIsLocking(fileToCheckPath), App.Logger);
 
 			if (processes is not null)
 			{
