+++ src/Files.App/Utils/Storage/Search/FolderSearch.cs
@@ -6,9 +6,7 @@ using System.IO;
 using Windows.Storage;
 using Windows.Storage.FileProperties;
 using Windows.Storage.Search;
-using static Files.Core.Helpers.NativeFindStorageItemHelper;
 using FileAttributes = System.IO.FileAttributes;
-using WIN32_FIND_DATA = Files.Core.Helpers.NativeFindStorageItemHelper.WIN32_FIND_DATA;
 
 namespace Files.App.Utils.Storage
 {
@@ -202,11 +200,11 @@ namespace Files.App.Utils.Storage
 
 			foreach (var match in matches)
 			{
-				(IntPtr hFile, WIN32_FIND_DATA findData) = await Task.Run(() =>
+				(IntPtr hFile, Win32PInvoke.WIN32_FIND_DATA findData) = await Task.Run(() =>
 				{
-					int additionalFlags = FIND_FIRST_EX_LARGE_FETCH;
-					IntPtr hFileTsk = FindFirstFileExFromApp(match.FilePath, FINDEX_INFO_LEVELS.FindExInfoBasic,
-						out WIN32_FIND_DATA findDataTsk, FINDEX_SEARCH_OPS.FindExSearchNameMatch, IntPtr.Zero, additionalFlags);
+					int additionalFlags = Win32PInvoke.FIND_FIRST_EX_LARGE_FETCH;
+					IntPtr hFileTsk = Win32PInvoke.FindFirstFileExFromApp(match.FilePath, Win32PInvoke.FINDEX_INFO_LEVELS.FindExInfoBasic,
+						out Win32PInvoke.WIN32_FIND_DATA findDataTsk, Win32PInvoke.FINDEX_SEARCH_OPS.FindExSearchNameMatch, IntPtr.Zero, additionalFlags);
 					return (hFileTsk, findDataTsk);
 				}).WithTimeoutAsync(TimeSpan.FromSeconds(5));
 
@@ -230,7 +228,7 @@ namespace Files.App.Utils.Storage
 						}
 					}
 
-					FindClose(hFile);
+					Win32PInvoke.FindClose(hFile);
 				}
 				else
 				{
@@ -284,11 +282,11 @@ namespace Files.App.Utils.Storage
 		private async Task SearchWithWin32Async(string folder, bool hiddenOnly, uint maxItemCount, IList<ListedItem> results, CancellationToken token)
 		{
 			//var sampler = new IntervalSampler(500);
-			(IntPtr hFile, WIN32_FIND_DATA findData) = await Task.Run(() =>
+			(IntPtr hFile, Win32PInvoke.WIN32_FIND_DATA findData) = await Task.Run(() =>
 			{
-				int additionalFlags = FIND_FIRST_EX_LARGE_FETCH;
-				IntPtr hFileTsk = FindFirstFileExFromApp($"{folder}\\{QueryWithWildcard}", FINDEX_INFO_LEVELS.FindExInfoBasic,
-					out WIN32_FIND_DATA findDataTsk, FINDEX_SEARCH_OPS.FindExSearchNameMatch, IntPtr.Zero, additionalFlags);
+				int additionalFlags = Win32PInvoke.FIND_FIRST_EX_LARGE_FETCH;
+				IntPtr hFileTsk = Win32PInvoke.FindFirstFileExFromApp($"{folder}\\{QueryWithWildcard}", Win32PInvoke.FINDEX_INFO_LEVELS.FindExInfoBasic,
+					out Win32PInvoke.WIN32_FIND_DATA findDataTsk, Win32PInvoke.FINDEX_SEARCH_OPS.FindExSearchNameMatch, IntPtr.Zero, additionalFlags);
 				return (hFileTsk, findDataTsk);
 			}).WithTimeoutAsync(TimeSpan.FromSeconds(5));
 
@@ -333,15 +331,15 @@ namespace Files.App.Utils.Storage
 							SearchTick?.Invoke(this, EventArgs.Empty);
 						}
 
-						hasNextFile = FindNextFile(hFile, out findData);
+						hasNextFile = Win32PInvoke.FindNextFile(hFile, out findData);
 					} while (hasNextFile);
 
-					FindClose(hFile);
+					Win32PInvoke.FindClose(hFile);
 				});
 			}
 		}
 
-		private ListedItem GetListedItemAsync(string itemPath, WIN32_FIND_DATA findData)
+		private ListedItem GetListedItemAsync(string itemPath, Win32PInvoke.WIN32_FIND_DATA findData)
 		{
 			ListedItem listedItem = null;
 			var isHidden = ((FileAttributes)findData.dwFileAttributes & FileAttributes.Hidden) == FileAttributes.Hidden;
