+++ src/Files.App/Utils/Storage/StorageItems/ZipStorageFile.cs
@@ -104,7 +104,7 @@ namespace Files.App.Utils.Storage
 						return await backingFile.OpenAsync(accessMode);
 					}
 
-					var file = NativeFileOperationsHelper.OpenFileForRead(containerPath, rw);
+					var file = Win32PInvoke.OpenFileForRead(containerPath, rw);
 					return file.IsInvalid ? null : new FileStream(file, rw ? FileAccess.ReadWrite : FileAccess.Read).AsRandomAccessStream();
 				}
 
@@ -149,7 +149,7 @@ namespace Files.App.Utils.Storage
 						return await backingFile.OpenReadAsync();
 					}
 
-					var hFile = NativeFileOperationsHelper.OpenFileForRead(containerPath);
+					var hFile = Win32PInvoke.OpenFileForRead(containerPath);
 					return hFile.IsInvalid ? null : new StreamWithContentType(new FileStream(hFile, FileAccess.Read).AsRandomAccessStream());
 				}
 
@@ -188,7 +188,7 @@ namespace Files.App.Utils.Storage
 						return await backingFile.OpenSequentialReadAsync();
 					}
 
-					var hFile = NativeFileOperationsHelper.OpenFileForRead(containerPath);
+					var hFile = Win32PInvoke.OpenFileForRead(containerPath);
 					return hFile.IsInvalid ? null : new FileStream(hFile, FileAccess.Read).AsInputStream();
 				}
 
@@ -310,7 +310,7 @@ namespace Files.App.Utils.Storage
 					else
 					{
 						var fileName = IO.Path.Combine(IO.Path.GetDirectoryName(Path), desiredName);
-						NativeFileOperationsHelper.MoveFileFromApp(Path, fileName);
+						Win32PInvoke.MoveFileFromApp(Path, fileName);
 					}
 				}
 				else
@@ -354,7 +354,7 @@ namespace Files.App.Utils.Storage
 					}
 					else if (option == StorageDeleteOption.PermanentDelete)
 					{
-						NativeFileOperationsHelper.DeleteFileFromApp(Path);
+						Win32PInvoke.DeleteFileFromApp(Path);
 					}
 					else
 					{
@@ -399,7 +399,7 @@ namespace Files.App.Utils.Storage
 		{
 			try
 			{
-				var hFile = NativeFileOperationsHelper.OpenFileForRead(path);
+				var hFile = Win32PInvoke.OpenFileForRead(path);
 				if (hFile.IsInvalid)
 				{
 					return false;
@@ -474,7 +474,7 @@ namespace Files.App.Utils.Storage
 				}
 				else
 				{
-					var hFile = NativeFileOperationsHelper.OpenFileForRead(containerPath, readWrite);
+					var hFile = Win32PInvoke.OpenFileForRead(containerPath, readWrite);
 					if (hFile.IsInvalid)
 					{
 						return null;
