+++ src/Files.App/Utils/Storage/StorageItems/ZipStorageFolder.cs
@@ -100,7 +100,7 @@ namespace Files.App.Utils.Storage
 		{
 			Func<Task<bool>> queryFileAssoc = async () =>
 			{
-				var assoc = await NativeWinApiHelper.GetFileAssociationAsync(filePath);
+				var assoc = await Win32PInvoke.GetFileAssociationAsync(filePath);
 				if (assoc is not null)
 				{
 					return assoc == Package.Current.Id.FamilyName
@@ -350,7 +350,7 @@ namespace Files.App.Utils.Storage
 					else
 					{
 						var fileName = IO.Path.Combine(IO.Path.GetDirectoryName(Path), desiredName);
-						NativeFileOperationsHelper.MoveFileFromApp(Path, fileName);
+						Win32PInvoke.MoveFileFromApp(Path, fileName);
 					}
 				}
 				else
@@ -397,7 +397,7 @@ namespace Files.App.Utils.Storage
 					}
 					else if (option == StorageDeleteOption.PermanentDelete)
 					{
-						NativeFileOperationsHelper.DeleteFileFromApp(Path);
+						Win32PInvoke.DeleteFileFromApp(Path);
 					}
 					else
 					{
@@ -488,7 +488,7 @@ namespace Files.App.Utils.Storage
 		{
 			return SafetyExtensions.IgnoreExceptions(() =>
 			{
-				var hFile = NativeFileOperationsHelper.OpenFileForRead(path);
+				var hFile = Win32PInvoke.OpenFileForRead(path);
 				if (hFile.IsInvalid)
 				{
 					return false;
@@ -529,7 +529,7 @@ namespace Files.App.Utils.Storage
 		{
 			return SafetyExtensions.IgnoreExceptions(() =>
 			{
-				var hFile = NativeFileOperationsHelper.OpenFileForRead(path, true);
+				var hFile = Win32PInvoke.OpenFileForRead(path, true);
 				if (hFile.IsInvalid)
 				{
 					return Task.FromResult(false);
@@ -580,7 +580,7 @@ namespace Files.App.Utils.Storage
 				}
 				else
 				{
-					var hFile = NativeFileOperationsHelper.OpenFileForRead(containerPath, readWrite);
+					var hFile = Win32PInvoke.OpenFileForRead(containerPath, readWrite);
 					if (hFile.IsInvalid)
 					{
 						return null;
