+++ src/Files.App/ViewModels/Properties/Items/BaseProperties.cs
@@ -4,7 +4,6 @@
 using Microsoft.UI.Dispatching;
 using System.IO;
 using Windows.Storage.FileProperties;
-using static Files.Core.Helpers.NativeFindStorageItemHelper;
 using FileAttributes = System.IO.FileAttributes;
 
 namespace Files.App.ViewModels.Properties
@@ -49,14 +48,14 @@ namespace Files.App.ViewModels.Properties
 
 			long size = 0;
 			long sizeOnDisk = 0;
-			FINDEX_INFO_LEVELS findInfoLevel = FINDEX_INFO_LEVELS.FindExInfoBasic;
-			int additionalFlags = FIND_FIRST_EX_LARGE_FETCH;
+			Win32PInvoke.FINDEX_INFO_LEVELS findInfoLevel = Win32PInvoke.FINDEX_INFO_LEVELS.FindExInfoBasic;
+			int additionalFlags = Win32PInvoke.FIND_FIRST_EX_LARGE_FETCH;
 
-			IntPtr hFile = FindFirstFileExFromApp(
+			IntPtr hFile = Win32PInvoke.FindFirstFileExFromApp(
 				path + "\\*.*",
 				findInfoLevel,
-				out WIN32_FIND_DATA findData,
-				FINDEX_SEARCH_OPS.FindExSearchNameMatch,
+				out Win32PInvoke.WIN32_FIND_DATA findData,
+				Win32PInvoke.FINDEX_SEARCH_OPS.FindExSearchNameMatch,
 				IntPtr.Zero,
 				additionalFlags);
 
@@ -72,7 +71,7 @@ namespace Files.App.ViewModels.Properties
 					if (((FileAttributes)findData.dwFileAttributes & FileAttributes.Directory) != FileAttributes.Directory)
 					{
 						size += findData.GetSize();
-						var fileSizeOnDisk = NativeFileOperationsHelper.GetFileSizeOnDisk(Path.Combine(path, findData.cFileName));
+						var fileSizeOnDisk = Win32PInvoke.GetFileSizeOnDisk(Path.Combine(path, findData.cFileName));
 						sizeOnDisk += fileSizeOnDisk ?? 0;
 						++count;
 						ViewModel.FilesCount++;
@@ -104,9 +103,9 @@ namespace Files.App.ViewModels.Properties
 					if (token.IsCancellationRequested)
 						break;
 				}
-				while (FindNextFile(hFile, out findData));
+				while (Win32PInvoke.FindNextFile(hFile, out findData));
 
-				FindClose(hFile);
+				Win32PInvoke.FindClose(hFile);
 
 				return (size, sizeOnDisk);
 			}
