+++ src/Files.App/Views/Properties/GeneralPage.xaml.cs
@@ -152,7 +152,7 @@ namespace Files.App.Views.Properties
 				}
 
 				if (ViewModel.IsUnblockFileSelected)
-					NativeFileOperationsHelper.DeleteFileFromApp($"{item.ItemPath}:Zone.Identifier");
+					Win32PInvoke.DeleteFileFromApp($"{item.ItemPath}:Zone.Identifier");
 
 				if (!GetNewName(out var newName))
 					return true;
