+++ src/Files.Core/Services/SizeProvider/CachedSizeProvider.cs
@@ -1,13 +1,8 @@
 ﻿// Copyright (c) 2023 Files Community
 // Licensed under the MIT License. See the LICENSE.
 
-using Files.Core.Extensions;
-using System;
 using System.Collections.Concurrent;
 using System.IO;
-using System.Threading;
-using System.Threading.Tasks;
-using static Files.Core.Helpers.NativeFindStorageItemHelper;
 
 namespace Files.Core.Services.SizeProvider
 {
